@@ -1,8 +1,8 @@
-module.exports = {
+export default {
   theme: {
     extend: {
       fontFamily: {
-        sans: "var(--font-inter)",
+        sans: 'var(--font-inter)',
       },
     },
   },
