@@ -19,7 +19,7 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
     <html lang="en" className={inter.variable}>
-      <body className="min-h-screen bg-white text-black font-sans">
+      <body className=" bg-white text-black font-sans">
         <AuthProvider>{children}</AuthProvider>
       </body>
     </html>
